extracted diagnostics testing mechanism, moved functions diagnostics to functions spec file
